Refactor services into dedicated modules
